Renaming to introduce new filtration type
