+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/internal/flag/ApparentPairsDetector.scala
@@ -1,4 +1,4 @@
-package io.github.jakipatryk.sparkpersistenthomology.internal.vr
+package io.github.jakipatryk.sparkpersistenthomology.internal.flag
 
 private[sparkpersistenthomology] object ApparentPairsDetector extends Serializable {
