+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/internal/flag/CliqueIterator.scala
@@ -1,4 +1,4 @@
-package io.github.jakipatryk.sparkpersistenthomology.internal.vr
+package io.github.jakipatryk.sparkpersistenthomology.internal.flag
 
 import io.github.jakipatryk.sparkpersistenthomology.internal.utils.CombinatorialNumberSystem
