+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/internal/flag/EnclosingRadiusCalculator.scala
@@ -1,4 +1,4 @@
-package io.github.jakipatryk.sparkpersistenthomology.internal.vr
+package io.github.jakipatryk.sparkpersistenthomology.internal.flag
 
 import org.apache.spark.sql.Dataset
 import org.apache.spark.broadcast.Broadcast
