+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/internal/flag/SimplexIndex.scala
@@ -1,4 +1,4 @@
-package io.github.jakipatryk.sparkpersistenthomology.internal.vr
+package io.github.jakipatryk.sparkpersistenthomology.internal.flag
 
 /** Represents a unique index of a simplex in the filtration.
   *
