+++ src/main/scala/io/github/jakipatryk/sparkpersistenthomology/internal/flag/SparseDistanceMatrix.scala
@@ -1,4 +1,4 @@
-package io.github.jakipatryk.sparkpersistenthomology.internal.vr
+package io.github.jakipatryk.sparkpersistenthomology.internal.flag
 
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.sql.SparkSession
