+++ src/test/scala/io/github/jakipatryk/sparkpersistenthomology/internal/flag/CliqueIteratorSpec.scala
@@ -1,4 +1,4 @@
-package io.github.jakipatryk.sparkpersistenthomology.internal.vr
+package io.github.jakipatryk.sparkpersistenthomology.internal.flag
 
 import io.github.jakipatryk.sparkpersistenthomology.internal.utils.CombinatorialNumberSystem
 import org.scalatest.flatspec.AnyFlatSpec
